fix codeql string_c_utils.c: 430

+++ src/utils/c_utils/src/string_c_utils.c
@@ -427,7 +427,7 @@ char* ADUC_StringUtils_Map(const char* src, int (*mapFn)(int))
         return NULL;
     }
 
-    for (int i = 0; i <= len; ++i)
+    for (size_t i = 0; i <= len; ++i)
     {
         int ret = mapFn(src[i]);
